# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "reprep API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/account/export": {
      "get": {
        "summary": "Export Account",
        "operationId": "export_account_api_account_export_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "schema_version": {
                      "type": "string"
                    },
                    "exported_at": {
                      "type": "string"
                    },
                    "account": {
                      "type": "object",
                      "description": "Own profile, including own MAX ID. No sessions, credentials or invitation hashes."
                    }
                  },
                  "required": [
                    "schema_version",
                    "exported_at",
                    "account"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "x-roles": [
          "tutor",
          "learner",
          "guardian"
        ],
        "security": [
          {
            "SessionBearer": []
          }
        ]
      }
    },
    "/api/account/deletion": {
      "get": {
        "summary": "Deletion Status",
        "operationId": "deletion_status_api_account_deletion_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "request": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "status": {
                              "type": "string"
                            },
                            "created": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "status",
                            "created"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "note": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "request",
                    "note"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "x-roles": [
          "tutor",
          "learner",
          "guardian"
        ],
        "security": [
          {
            "SessionBearer": []
          }
        ]
      },
      "post": {
        "summary": "Request Deletion",
        "operationId": "request_deletion_api_account_deletion_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "status"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Version, state or replay conflict when applicable"
          }
        },
        "x-roles": [
          "tutor",
          "learner",
          "guardian"
        ],
        "security": [
          {
            "SessionBearer": []
          }
        ]
      }
    },
    "/api/account/deletion/cancel": {
      "post": {
        "summary": "Cancel Deletion",
        "operationId": "cancel_deletion_api_account_deletion_cancel_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Version, state or replay conflict when applicable"
          }
        },
        "x-roles": [
          "tutor",
          "learner",
          "guardian"
        ],
        "security": [
          {
            "SessionBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeletionRequest": {
        "properties": {
          "confirm_alias": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Confirm Alias"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "confirm_alias"
        ],
        "title": "DeletionRequest"
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "reference_id": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "reference_id"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
    "securitySchemes": {
      "SessionBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "MaxWebhookSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Max-Bot-Api-Secret"
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local demo only; configure approved HTTPS for judging"
    }
  ]
}